+++ packages/build-scripts/test/configFile/esmConfig.js
@@ -1,0 +1,5 @@
+const config = {
+  entry: 'src/mjsFile.mjs',
+};
+
+export default config;