+++ packages/build-scripts/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.0
+
+- [refactor] support custom command by extend Context
+- [feat] support config written with typescript and es module
+- [feat] enhance API modifyUserConfig while modify userConfig by config path `modifyUserConfig('output.path', 'dir')`
+- [feat] support deep merge of modifyUserConfig by options
+- [feat] enhance registerMethod API, make it possible to get plugin name when applyMethod
+- [feat] add `originalUserConfig` to plugin API
+- [feat] support `hasRegistration` api
+- [fix] move webpack-dev-server to peerDependencies and migrate webpack-dev-server to 4.0.0
+
 ## 1.0.1
 
 - [chore] bump version because of 1.0.0 has been previously published